‎

+++ webroot/common_scripts/patch.sh
@@ -5,66 +5,46 @@ LOG="/data/adb/Integrity-Box-Logs"
 LOGFILE="$LOG/patch.log"
 TARGET_DIR="/data/adb/tricky_store"
 FILE_PATH="$TARGET_DIR/security_patch.txt"
-FILE_CONTENT="all=2025-05-05"
+FILE_CONTENT="all=2025-06-05"
 U="/data/adb/modules/integrity_box"
 
+# Execute kill script
 chmod +x "$U/kill.sh"
 sh "$U/kill.sh"
 
-# Ensure directories exist
+# Ensure log directory exists
 mkdir -p "$LOG"
-mkdir -p "$TARGET_DIR"
 
 # Logging function
 log() { echo -e "$1" | tee -a "$LOGFILE"; }
 
-# popup function
+# Popup function
 popup() { am start -a android.intent.action.MAIN -e mona "$@" -n meow.helper/.MainActivity >/dev/null 2>&1; sleep 0.5; }
 
-# Function to Detect Key Press
-CheckKey() {
-  while true; do
-    key=$(getevent -qlc 1 | awk '/KEY_/ {print $3; exit}')
-    case $key in
-      KEY_VOLUMEUP|KEY_VOLUMEDOWN|KEY_POWER) echo "$key"; return ;;
-    esac
-    sleep 0.1
-  done
-}
-
-# Start Logging
-log "\n━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
-log "  📝 Logged on:  $(date '+%A %d/%m/%Y %I:%M:%S %p')"
-log "━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
+log "\n━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
 log " "
+log "Patch Mode : Auto"
+log "Spoofed to : 05 MAY 2025"
+log "Applied on:  $(date '+%A %d/%m/%Y %I:%M:%S %p')"
 
-#: Security Patch Spoofing
-log "For A13+ checks only"
-log "\n📢 Security Patch Hack Options:"
-log "━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
-log "   [➕]  Spoof security patch"
-log "   [➖]  Remove spoofed patch"
-log "   [🔴]  Cancel operation"
-log "━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
-log " "
-popup " Vol+ Enable / Vol- Disable"
-log " "
+# Check if file exists to proceed
+if [ ! -f "$FILE_PATH" ] && [ ! -d "$TARGET_DIR" ]; then
+    log "⚠️ Tricky Store is not installed."
+    popup "⚠️ Tricky Store is not installed"
+    exit 0
+fi
 
-case $(CheckKey) in
-  KEY_VOLUMEUP) 
-    echo "$FILE_CONTENT" > "$FILE_PATH"
-    log "✅ Security patch spoofed successfully!"
-    popup "✅ Spoof applied!"
-    ;;
-  KEY_VOLUMEDOWN) 
+# Toggle spoofing
+if [ -f "$FILE_PATH" ]; then
     rm -f "$FILE_PATH"
-    log "🗑️ Spoof removed!"
-    popup "🗑️ File removed!"
-    ;;
-  KEY_POWER) 
-    log "❌ Operation canceled!"
-    popup "❌ Canceled!"
-    exit 0 ;;
-esac
+    log "Patch Status : ❌ Not Spoofed / Removed"
+    popup "🗑️ Spoof removed!"
+else
+    mkdir -p "$TARGET_DIR"
+    echo "$FILE_CONTENT" > "$FILE_PATH"
+    log "Patch Status : ✅ Spoofed"
+    popup "✅ Spoof applied!"
+fi
 
+log "\n━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
 log " "
